Implement multi-TTY output
The `-output` module subcommand allows to specify where a single module
will be written. If more modules share the same output they will be
stacked as usual.

This is a possibly breaking change for custom modules as it introduces
the `term_width` argument to the `lines` method and also changes the
signature of `divider`. This is needed because now each module may be
written to a possibly different terminal/file, a single global terminal
width is no more feasible. Custom modules must take into account the
aforementioned argument to generate the content.