Highlight changes to the memory regions that are being watched.
Configurable via:

   dashboard -style memory highlight-changes False|True|"cumulative"

False        means the feature is disabled.
True         means to highlight the changes of exactly the last gdb command.
"cumulative" means to highlight all changes, where changes of the last
             gdb command are highlighted specially.

The default is True.

The highlighted changes can be cleared by unwatching, and (re)watching
the particular memory region.

Tested with gdb 8.1.0, and Python 3.6.5.