Allow source and assembly modules to take the whole terminal height
By passing 0 to the height, those modules use the whole terminal space. Of
course it only does make sense when a module is the only one bound to a given
terminal.

This is also a breaking change for the module API as now the height too is
passed to the `lines` method.

Close #51.