Use compact formatting for values by default
The idea is to show all the modules by default, yet with the previous default a
user could easily end up with a *huge* dashboard due to a great number of
variables and/or complex data structures (e.g., C++). This change just makes the
above less likely. The same problem is currently still present with the thread
module though...