Buffer the dashboard content to avoid some flickering
This avoids the *internal* flickering when a module is slower compared to some
prior module but unfortunately (while stepping) the flickering of the dashboard
compared to the "Output/messages" content cannot be avoided. Currently the best
solution is to display the dashboard in a separate terminal while using the main
terminal for prompt and inferior I/O only.