Improve default color for gray
A couple of styles by default use 1;30. However 30 stands for black and
1 stands for bold or increased intensity. So depending on the
interpretation, 1;30 may simply result in black. For terminals with a
black or dark background that interpret 1 not as increased intensity,
the output text is therefore not or almost not readable.

This is the case for gvim or neovim (with termguicolors set) but also
for some colorschemes of konsole, and probably others. In the case of
gvim/neovim this impacts the :TermDebug use case.

Since the original intent of 1;30 was probably to obtain a gray, this
change sets all occurences of 1;30 to 0;90, which results in gray again,
but without relying on the interpretation of 1 as an high-intense
modifier.