rewrite memory module to use GDB's "x" output
This is a rewrite of the memory module to enable watching a memory
region with the format options of GDB's "x" command. Instead of
getting a raw memory dump like before, an "x" command is created,
evaluated by GDB and its output is formatted according to the
dashboard settings. The watch command of the previous memory
module can be used and extended for backwards compatibility.
Alternatively the watch command also accepts an "x" command as
it would be entered into GDB directly.