Fix address display for instructions and strings; minor cleanup
Instructions and strings have variable sizes, so the addresses for
consecutive objects can't be calculated. To fix this, display only one
object per line when the format is "i" or "s" (like in GDB's x output)
and keep the addresses from GDB's output. This also has the benefit of
displaying address tags.