Fix GDB tab completion issue
So it seems that importing readline causes issue with the GDB prompt, in fact
GDB already tries (failing) to avoid that. The issue manifests when debugging a
program with the source code available, e.g., C, and IPython is available in the
system.

Fixes #325 and closes #326 by providing a better solution.